Added boost thread pool wrapper class to replace my implementation
